Improvements to various Effects and UI stuff
- better delays and alignment of damage numbers, more accurate mob hp
bar
- added scrolling server notice
- added two handlers for different server notices
- added util for interpolation to improve consistency

+++ Net/Packets/AttackAndSkillPackets.h
@@ -104,11 +104,11 @@ namespace Net
 	class UseSkillPacket : public OutPacket
 	{
 	public:
-		UseSkillPacket(int32_t skillid, uint8_t level) : OutPacket(USE_SKILL)
+		UseSkillPacket(int32_t skillid, int32_t level) : OutPacket(USE_SKILL)
 		{
 			writetime();
 			writeint(skillid);
-			writech(level);
+			writech(static_cast<uint8_t>(level));
 
 			// if monster magnet : some more bytes
